requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
html2text>=2020.1.16
readability-lxml>=0.8.1

# DeepSeek API (using OpenAI SDK)
openai>=1.50.0

# Image Processing
Pillow>=10.2.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
validators>=0.22.0
tenacity>=8.2.3
loguru>=0.7.2
